feat: Implement Saving Result Utils
Until now, I manually save log and check result in log files, but things are changed.
I do not need to save result anymore and it is not the end.
Here I summarized what features are added today.
- auto save result in csv format.
- auto save best result in csv format
- auto mark best result calculate by averaging 3 scores
- auto delete too old logs, tensorboard, weights
So you can focus on the research from now on, isn't it great?